Update Simulation
The libraries were necessary to load so that I could import my data into a gif. Then, I added the data of steps I take in a day over 14 days. By putting c(data, data, ...) It added this information to the y axis that I will eventually specify in the code. (i in 1:4) gives my days over 2 weeks individual frames. Then, I specified in the code that days (time) goes on the x axis, and my individual data on the y axis. I then creating my plot and imported it into a GIF.